requirements.txt
Adafruit-Blinka==8.66.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-framebuf==1.6.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-ssd1306==2.12.21
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.83.1
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.10.0
args==0.1.0
binho-host-adapter==0.1.6
clint==0.5.1
htu21df==0.1.4
lgpio==0.2.2.0
mock==5.2.0
paho-mqtt==2.1.0
pillow==11.3.0
pyftdi==0.57.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.17.0
smbus==1.1.post2
sysv-ipc==1.1.0
typing_extensions==4.15.0
wrapt==1.17.3
